chore: convert settings and licenses to CMP and MVI; navigate using Circuit from settings to licenses
