+++ shared/feature/settings/src/androidMain/kotlin/ly/david/musicsearch/shared/feature/settings/internal/components/SettingWithDialogChoicesPreview.kt
@@ -1,0 +1,21 @@
+package ly.david.musicsearch.shared.feature.settings.internal.components
+
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import ly.david.musicsearch.shared.feature.settings.internal.components.SettingWithDialogChoices
+import ly.david.ui.core.preview.DefaultPreviews
+import ly.david.ui.core.theme.PreviewTheme
+
+@DefaultPreviews
+@Composable
+internal fun PreviewSettingWithDialogChoices() {
+    PreviewTheme {
+        Surface {
+            SettingWithDialogChoices(
+                title = "Theme",
+                choices = listOf("Light"),
+                selectedChoiceIndex = 0,
+            )
+        }
+    }
+}
