+++ shared/feature/settings/src/androidMain/kotlin/ly/david/musicsearch/shared/feature/settings/internal/licenses/Licenses.kt
@@ -1,0 +1,24 @@
+package ly.david.musicsearch.shared.feature.settings.internal.licenses
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.mikepenz.aboutlibraries.ui.compose.LibrariesContainer
+import com.mikepenz.aboutlibraries.ui.compose.LibraryDefaults
+
+@Composable
+internal actual fun Licenses(
+    modifier: Modifier,
+) {
+    LibrariesContainer(
+        modifier = Modifier
+            .fillMaxSize(),
+        colors = LibraryDefaults.libraryColors(
+            backgroundColor = MaterialTheme.colorScheme.background,
+            contentColor = MaterialTheme.colorScheme.onBackground,
+            badgeBackgroundColor = MaterialTheme.colorScheme.primary,
+            badgeContentColor = MaterialTheme.colorScheme.onPrimary,
+        ),
+    )
+}
