+++ shared/feature/settings/src/androidUnitTest/kotlin/ly/david/musicsearch/shared/feature/settings/components/SettingSwitchTest.kt
@@ -1,0 +1,26 @@
+package ly.david.musicsearch.shared.feature.settings.components
+
+import com.google.testing.junit.testparameterinjector.TestParameterInjector
+import ly.david.musicsearch.shared.feature.settings.internal.components.PreviewSettingSwitchChecked
+import ly.david.musicsearch.shared.feature.settings.internal.components.PreviewSettingSwitchUnchecked
+import ly.david.ui.test.screenshot.ScreenshotTest
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(TestParameterInjector::class)
+class SettingSwitchTest : ScreenshotTest() {
+
+    @Test
+    fun checked() {
+        snapshot {
+            PreviewSettingSwitchChecked()
+        }
+    }
+
+    @Test
+    fun unchecked() {
+        snapshot {
+            PreviewSettingSwitchUnchecked()
+        }
+    }
+}
