+++ shared/feature/settings/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/settings/internal/DevSettingsSection.kt
@@ -1,0 +1,12 @@
+package ly.david.musicsearch.shared.feature.settings.internal
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.runtime.Composable
+import ly.david.ui.common.listitem.ListSeparatorHeader
+
+@Composable
+internal fun DevSettingsSection() {
+    Column {
+        ListSeparatorHeader(text = "Dev Settings")
+    }
+}
