+++ shared/feature/settings/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/settings/internal/SettingsUiState.kt
@@ -1,0 +1,14 @@
+package ly.david.musicsearch.shared.feature.settings.internal
+
+import com.slack.circuit.runtime.CircuitUiState
+import ly.david.musicsearch.core.preferences.AppPreferences
+
+internal data class SettingsUiState(
+    val username: String,
+    val accessToken: String?,
+    val theme: AppPreferences.Theme,
+    val useMaterialYou: Boolean,
+    val showMoreInfoInReleaseListItem: Boolean = true,
+    val sortReleaseGroupListItems: Boolean = false,
+    val eventSink: (SettingsUiEvent) -> Unit,
+) : CircuitUiState
