+++ shared/feature/settings/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/settings/internal/licenses/Licenses.kt
@@ -1,0 +1,9 @@
+package ly.david.musicsearch.shared.feature.settings.internal.licenses
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+internal expect fun Licenses(
+    modifier: Modifier,
+)
