--- shared/feature/settings/src/main/kotlin/ly/david/musicsearch/shared/feature/settings/DevSettingsSection.kt
@@ -1,12 +1,0 @@
-package ly.david.musicsearch.shared.feature.settings
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.runtime.Composable
-import ly.david.ui.common.listitem.ListSeparatorHeader
-
-@Composable
-internal fun DevSettingsSection() {
-    Column {
-        ListSeparatorHeader(text = "Dev Settings")
-    }
-}
