--- shared/feature/settings/src/main/kotlin/ly/david/musicsearch/shared/feature/settings/SettingsScaffold.kt
@@ -1,80 +1,0 @@
-package ly.david.musicsearch.shared.feature.settings
-
-import android.content.Context
-import android.content.Intent
-import android.provider.Settings
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Scaffold
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.core.app.NotificationManagerCompat
-import ly.david.musicsearch.core.models.navigation.Destination
-import ly.david.musicsearch.core.preferences.AppPreferences
-import ly.david.musicsearch.strings.LocalStrings
-import ly.david.ui.common.topappbar.ScrollableTopAppBar
-import org.koin.androidx.compose.koinViewModel
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun SettingsScaffold(
-    modifier: Modifier = Modifier,
-    onDestinationClick: (Destination) -> Unit = {},
-    onLoginClick: () -> Unit = {},
-    onLogoutClick: () -> Unit = {},
-    showMoreInfoInReleaseListItem: Boolean = true,
-    onShowMoreInfoInReleaseListItemChange: (Boolean) -> Unit = {},
-    sortReleaseGroupListItems: Boolean = false,
-    onSortReleaseGroupListItemsChange: (Boolean) -> Unit = {},
-    viewModel: SettingsViewModel = koinViewModel(),
-) {
-    val strings = LocalStrings.current
-
-    Scaffold(
-        modifier = modifier,
-        topBar = {
-            ScrollableTopAppBar(
-                showBackButton = false,
-                title = strings.settings,
-            )
-        },
-    ) { innerPadding ->
-
-        val context = LocalContext.current
-
-        val username by viewModel.musicBrainzAuthStore.username.collectAsState(initial = "")
-        val accessToken by viewModel.musicBrainzAuthStore.accessToken.collectAsState(initial = null)
-        val theme by viewModel.appPreferences.theme.collectAsState(initial = AppPreferences.Theme.SYSTEM)
-        val useMaterialYou by viewModel.appPreferences.useMaterialYou.collectAsState(initial = true)
-
-        SettingsScreen(
-            modifier = Modifier.padding(innerPadding),
-            username = username,
-            showLogin = accessToken.isNullOrEmpty(),
-            onLoginClick = onLoginClick,
-            onLogoutClick = onLogoutClick,
-            onDestinationClick = onDestinationClick,
-            theme = theme,
-            onThemeChange = { viewModel.appPreferences.setTheme(it) },
-            useMaterialYou = useMaterialYou,
-            onUseMaterialYouChange = { viewModel.appPreferences.setUseMaterialYou(it) },
-            showMoreInfoInReleaseListItem = showMoreInfoInReleaseListItem,
-            onShowMoreInfoInReleaseListItemChange = onShowMoreInfoInReleaseListItemChange,
-            sortReleaseGroupListItems = sortReleaseGroupListItems,
-            onSortReleaseGroupListItemsChange = onSortReleaseGroupListItemsChange,
-            isNotificationListenerEnabled = context.isNotificationListenerEnabled(),
-            onGoToNotificationListenerSettings = {
-                context.startActivity(Intent(Settings.ACTION_NOTIFICATION_LISTENER_SETTINGS))
-            },
-            versionName = BuildConfig.VERSION_NAME,
-            versionCode = BuildConfig.VERSION_CODE,
-        )
-    }
-}
-
-private fun Context.isNotificationListenerEnabled(): Boolean {
-    return NotificationManagerCompat.getEnabledListenerPackages(this).any { it == this.packageName }
-}
