--- shared/feature/settings/src/main/kotlin/ly/david/musicsearch/shared/feature/settings/SettingsUiModule.kt
@@ -1,8 +1,0 @@
-package ly.david.musicsearch.shared.feature.settings
-
-import org.koin.core.annotation.ComponentScan
-import org.koin.core.annotation.Module
-
-@Module
-@ComponentScan
-class SettingsUiModule
