--- shared/feature/settings/src/main/kotlin/ly/david/musicsearch/shared/feature/settings/components/SettingSwitch.kt
@@ -1,83 +1,0 @@
-package ly.david.musicsearch.shared.feature.settings.components
-
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.selection.toggleable
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Switch
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.semantics.clearAndSetSemantics
-import androidx.compose.ui.unit.dp
-import ly.david.ui.core.preview.DefaultPreviews
-import ly.david.ui.core.theme.PreviewTheme
-import ly.david.ui.core.theme.TextStyles
-
-@Composable
-internal fun SettingSwitch(
-    header: String,
-    checked: Boolean,
-    modifier: Modifier = Modifier,
-    onCheckedChange: (Boolean) -> Unit = {},
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .toggleable(
-                value = checked,
-                role = Role.Switch,
-                onValueChange = onCheckedChange,
-            )
-            .padding(
-                horizontal = 16.dp,
-                vertical = 8.dp,
-            ),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = header,
-            style = TextStyles.getCardBodyTextStyle(),
-        )
-
-        Spacer(modifier = Modifier.weight(1f))
-
-        Switch(
-            checked = checked,
-            onCheckedChange = onCheckedChange,
-            modifier = Modifier.clearAndSetSemantics {},
-        )
-    }
-}
-
-// region Previews
-@DefaultPreviews
-@Composable
-internal fun PreviewSettingSwitchChecked() {
-    PreviewTheme {
-        Surface {
-            SettingSwitch(
-                header = "A setting",
-                checked = true,
-            )
-        }
-    }
-}
-
-@DefaultPreviews
-@Composable
-internal fun PreviewSettingSwitchUnchecked() {
-    PreviewTheme {
-        Surface {
-            SettingSwitch(
-                header = "A setting",
-                checked = false,
-            )
-        }
-    }
-}
-// endregion
