--- shared/feature/settings/src/test/kotlin/ly/david/musicsearch/shared/feature/settings/SettingsScreenTest.kt
@@ -1,24 +1,0 @@
-package ly.david.musicsearch.shared.feature.settings
-
-import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import ly.david.ui.test.screenshot.ScreenshotTest
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(TestParameterInjector::class)
-class SettingsScreenTest : ScreenshotTest(isFullScreen = true) {
-
-    @Test
-    fun default() {
-        snapshot {
-            PreviewSettingsScreen()
-        }
-    }
-
-    @Test
-    fun notificationListenerEnable() {
-        snapshot {
-            PreviewSettingsScreenNotificationListenerEnable()
-        }
-    }
-}
